% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.R
\name{digit.fixed2}
\alias{digit.fixed2}
\title{Digitize 3D landmarks on mesh3d object}
\usage{
digit.fixed2(spec = spec, out.dir = NULL)
}
\arguments{
\item{spec}{a list, produced from \code{read.ply2}}

\item{out.dir}{character, the file path describing where to save results}
}
\value{
A csv file with x,y,z coordinates named according to the \code{spec.name} output of \code{read.ply2}
}
\description{
A \code{shiny}-based interactive function to digitize three-dimensional (3D) landmarks. Input for the function is output of vertex coordinates defining a mesh3d object as obtained from \code{read.ply2}.
}
\details{
Function for digitizing fixed three-dimensional landmarks. Produces a \code{shiny} app interactive session. Permits panning, zooming, and rotation. Results are saved as a .csv file once the user clicks on the "Save" button.
}
\examples{

\dontrun{
#load a .ply file, taken from https://people.sc.fsu.edu/~jburkardt/data/ply/ply.html

ply.f<- system.file("extdata","hammerhead.ply",package = "geomorphcompanion")

spec <- read.ply2(file=ply.f,ShowSpecimen = F)

digit.fixed2(spec=spec,out.dir = tempdir())

#commence digitizing

f <- list.files(tempdir(),pattern=".csv",full.names=TRUE)
read_csv(f)
}
}
\seealso{
\code{geomorph::digitize.fixed}
}
