% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{readland.from1.nts}
\alias{readland.from1.nts}
\title{Read single *.nts file containing landmark coordinates for a single specimen, avoiding the bugs of \code{geomorph::readland.nts}}
\usage{
readland.from1.nts(file)
}
\arguments{
\item{file}{Path to the input file}
}
\value{
Modified from \code{geomorph::readland.nts} . . . Function returns a 3D array (p x k x n=1), where p is the number of landmark points, k is the number of landmark dimensions (2 or 3), and n is the number of specimens (one in this limited case). The third dimension of this array contains a name for the specimen which is obtained from the names in the *.nts file.
}
\description{
A function modified from \code{geomorph::readland.nts} to take a single .nts file and return an array that can be read by the \code{geomorph} package.
}
\examples{

#load a single nts file

file <- system.file("extdata", "echiostoma.nts", package = "geomorphcompanion")
nts.file <- readland.from1.nts(file=file)
print(nts.file)
}
\seealso{
\code{geomorph::readland.nts}
}
