% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{readmulti.from1.nts}
\alias{readmulti.from1.nts}
\title{Read multiple nts files.}
\usage{
readmulti.from1.nts(filelist)
}
\arguments{
\item{filelist}{A vector containing the file paths to all the .nts files to be compiled}
}
\value{
Modified from \code{geomorph::readmulit.nts} . . . Function returns a 3D array (p x k x n), where p is the number of landmark points, k is the number of landmark dimensions (2 or 3), and n is the number of specimens. The third dimension of this array contains names for each specimen, which are retrieved from the nts specimen labels, if those are available.
}
\description{
Read multiple nts (or .dta) files that reflect one specimen to obtain landmark coordinates and combine them into a single array. Avoiding the bugs of \code{geomorph::readland.nts}, about which \code{geomorph::readlmulti.nts} is wrapped
}
\details{
This is a wrapper of \link{readland.from1.nts} to allow reading of landmark coordinates, in 2D or 3D, from several .nts (or .dta) files each containing one specimen, and compiling them into an array for proceeding with \code{geomorph} procedures.
}
\examples{

library(geomorph)
#load a two nts files

file1 <- system.file("extdata","echiostoma.nts",package = "geomorphcompanion")
file2 <- system.file("extdata","echiostoma2.nts",package = "geomorphcompanion")
files <- c(file1,file2)
nts.file <- readmulti.from1.nts(filelist=files)
print(nts.file)

#perform GPA and plot
gpa <- gpagen(nts.file)
plot(gpa)
}
\seealso{
\code{geomorph::readmulti.nts}
}
